start implementation of booking facade

+++ src/main/java/epamUniversity/services/BookingService.java
@@ -30,7 +30,7 @@ public interface BookingService {
      * @return total price
      */
     public double getTicketsPrice(@Nonnull Event event, @Nonnull DateTime dateTime, @Nullable User user,
-                                  @Nonnull Set<Long> seats);
+                                  @Nonnull Set<Integer> seats);
 
     /**
      * Books tickets in internal system. If user is not
